Апка/tests/requirements.txt
pytest==8.4.1
pytest-timeout==2.4.0
pytest-asyncio==0.24.0
pytest-playwright==0.5.2
httpx==0.28.1
allure-pytest==2.14.3
playwright==1.55.0
pydantic==2.11.7
jsonschema==4.25.1
psycopg[binary]==3.2.1
SQLAlchemy
asyncpg

python-jose
passlib[bcrypt]
bcrypt==4.1.3
fastapi
aio-pika
python-multipart
jinja2
itsdangerous
pydantic-settings
